requirements.txt
streamlit==1.29.0
finnhub-python==2.4.21
pandas==2.1.4
plotly==5.18.0 